+++ server/schemas/Protein.js
@@ -1,10 +1,10 @@
-const { Text, Float, Checkbox } = require("@keystonejs/fields");
+const { Text, Integer, Checkbox } = require("@keystonejs/fields");
 
 module.exports = {
   fields: {
     name: { type: Text },
-    price: { type: Float },
-    stock: { type: Float },
+    price: { type: Integer },
+    stock: { type: Integer },
     vegitarian: { type: Checkbox },
   },
 };
